+++ common/src/main/java/io/github/mortuusars/envelope/client/gui/screen/PaybackPackageScreen.java
@@ -1,11 +1,9 @@
 package io.github.mortuusars.envelope.client.gui.screen;
 
 import io.github.mortuusars.envelope.Envelope;
-import io.github.mortuusars.envelope.client.util.Minecrft;
-import io.github.mortuusars.envelope.world.GameTime;
 import io.github.mortuusars.envelope.world.inventory.PaybackPackageMenu;
 import io.github.mortuusars.envelope.world.inventory.slot.PreviewSlot;
-import io.github.mortuusars.envelope.world.item.component.mail.log.DeliveryRecord;
+import io.github.mortuusars.envelope.world.item.PaybackPackageItem;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
@@ -43,10 +41,7 @@ public class PaybackPackageScreen extends AbstractInHandContainerScreen<PaybackP
         List<Component> components = super.getTooltipFromContainerItem(stack);
 
         if (hoveredSlot instanceof PreviewSlot) {
-            components.add(Component.literal("⌛ ")
-                  .append(GameTime.formatLargest(
-                        getMenu().getPaybackSubject().timeoutTick() - Minecrft.level().getGameTime(), false))
-                  .withStyle(DeliveryRecord.MessageType.NEGATIVE.getStyle()));
+            PaybackPackageItem.appendPaybackSubjectHoverText(components, getMenu().getPaybackSubject());
         }
 
         return components;